pyproject.toml
[project]
name = "f1api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.13.2",
    "fastapi==0.115.2",
    "psycopg[binary]==3.2.3",
    "pydantic-settings==2.4.0",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.30.6",
]

[dependency-groups]
dev = [
    "black==24.8.0",
    "httpx==0.27.2",
    "mypy==1.11.1",
    "pre-commit==3.8.0",
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "ruff==0.5.7",
]

[tool.mypy]
python_version = "3.13"
warn_unused_ignores = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = []

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "B", "I"]  # style, errors, bugbear, import sorting
lint.fixable = ["ALL"]

[tool.pytest.ini_options]
addopts = "-q"
